% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_constructor.R
\name{.is_datatype}
\alias{.is_datatype}
\title{Check object multi-omics data type - specific to MAP}
\usage{
.is_datatype(datatype)
}
\description{
Check object multi-omics data type - specific to MAP
}
