% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data_ids.R
\name{get_all_data_ids}
\alias{get_all_data_ids}
\title{Get all existing object IDs}
\usage{
get_all_data_ids(con, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}
}
\value{
a character vector of IDs of existing data
}
\description{
Get all existing object IDs
}
\details{
If the SHINYPROXY_USERNAME environment variable is set this will
return all IDs for that user, otherwise it will return all IDs in the
data connection.
}
