% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve data object from a moap_data_connection object using a UUID}
\usage{
get_data(con, id, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}

\item{id}{data object ID, typically returned by \code{\link{put_data}}}
}
\value{
R object stored at the specified ID
}
\description{
Retrieve data object from a moap_data_connection object using a UUID
}
\seealso{
\code{\link{put_data}}, \code{\link{map_data_connection}}
}
