% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Retrieve file object from a map_data_connection object using a UUID}
\usage{
get_file(con, id, filename = NULL, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}

\item{id}{file object ID, typically returned by \code{\link{put_file}}}

\item{filename}{(optional) local path to store object, if none provided, object is stored in a temp file.}
}
\value{
path to local file copy (\code{filename} if specified, otherwise generated with \code{\link{tempfile}} function)
}
\description{
Retrieve file object from a map_data_connection object using a UUID
}
\seealso{
\code{\link{put_file}}, \code{\link{map_data_connection}}
}
