% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tags associated with an ID}
\usage{
get_tags(con, id)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object using Min.io as the backend}

\item{id}{data object ID, typically returned by \code{\link{put_data}}}
}
\value{
a list object with key/value pairs
}
\description{
Get tags associated with an ID
}
\details{
Only valid when using Min.io as the backend storage mechanism. Tags are not
guaranteed to be in the same order as provided in {\link{set_tags}}.
}
\seealso{
\code{\link{set_tags}}
}
