% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetype_checks.R
\name{is_emeta}
\alias{is_emeta}
\title{Test if a file is an e_meta file. Its companion edata file must be provided.}
\usage{
is_emeta(edata, emeta)
}
\arguments{
\item{edata}{Must be a dataframe or datatable. Required.}

\item{emeta}{Must be a dataframe or datatable. Required.}
}
\value{
A boolean where TRUE means the file is an acceptable e_meta file.
}
\description{
Test if a file is an e_meta file. Its companion edata file must be provided.
}
\examples{
\dontrun{

library(pmartRdata)
is_emeta(edata = pmartRdata::pep_edata, emeta = pmartRdata::pep_emeta)

}
}
