% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetype_checks.R
\name{is_fdata}
\alias{is_fdata}
\title{Test if a file is an f_data file. Its companion edata file must be provided.}
\usage{
is_fdata(edata, fdata)
}
\arguments{
\item{edata}{Must be a dataframe or datatable. Required.}

\item{fdata}{Must be a dataframe or datatable. Required.}
}
\value{
A boolean where TRUE means the file is an acceptable f_data file.
}
\description{
Test if a file is an f_data file. Its companion edata file must be provided.
}
\examples{
\dontrun{

library(pmartRdata)
is_fdata(edata = pmartRdata::metab_edata, fdata = pmartRdata::metab_fdata)

}
}
