% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data_connection.R
\name{map_data_connection}
\alias{map_data_connection}
\title{Construct data connection object from config file or environment variables}
\usage{
map_data_connection(config_file = NA)
}
\arguments{
\item{config_file}{optional config file}
}
\value{

}
\description{
Construct data connection object from config file or environment variables
}
\examples{
\dontrun{
con <- map_data_connection(config_file=system.file("example_directory_config.yml", package="mapDataAccess"))
id <- put_data(con, iris)
iris2 <- get_data(con, id)
identical(iris, iris2)
}
}
