% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint_constructor.R
\name{midpoint_ipmart}
\alias{midpoint_ipmart}
\title{Generate a midpoint analysis file for ipmart}
\usage{
midpoint_ipmart(midpoints, tab, name = "exportedFromIPMART", fmeta = NULL)
}
\arguments{
\item{midpoints}{List of midpoint pmart objects from the same tab. Must contain 2-5 objects. There can be
no more than 2 metabolomics (1 of: NMR or GC/LC-MS), no more than 2 lipidomics datasets,
and no more than 1 proteomics (peptide or protein) dataset. Required.}

\item{tab}{The ipmart tab the data was exported from. Acceptable entries are
"normalization_tab", "statistics_tab", or "integration_tab". Required.}

\item{fmeta}{Must be a dataframe or data table. If not provided, users can built it in
ipmart. Default is NULL.}
}
\value{
A midpoint ipmart object
}
\description{
Possible "exit" points to generate an midpoint file in ipmart are
normalization, peptide statistics, peptide rollup, and statistics.
}
\examples{
\dontrun{

# Generate the two midpoint pmart objects from in the "midpoint_pmart" example

midpoint_ipmart(midpoints = list(pep_midpoint, lipid_midpoint),
                tab = "normalization_tab", 
                fmeta = NULL)

}
}
