% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_constructor.R
\name{project_edata}
\alias{project_edata}
\title{Generate a simple e_data only project that can only be opened in MODE}
\usage{
project_edata(projectname, datatype, edata, edata_filename = NULL)
}
\arguments{
\item{projectname}{Any string to name the project. All spaces and non-alphanumeric
characters will be removed to prevent issues with the visualizations. Required.}

\item{datatype}{Must be of the acceptable MAP omic class. See .is_datatype for list. Required.}

\item{edata}{Must be a dataframe or datatable. Required.}

\item{edata_name}{The path or name for the edata file. Optional.}
}
\value{
A project edata object
}
\description{
Construct a project edata object where only edata is required.
}
\examples{
\dontrun{

library(pmartRdata)
project_edata(projectname = "LipidToMode",
             datatype = "Lipidomics-Positive",
             edata = pmartRdata::lipid_edata_pos,
             edata_filename = "lip_edata")

}
}
