% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_constructor.R
\name{project_multiomics}
\alias{project_multiomics}
\title{Generate a multi-omics project object to pass data from MAP to IPMART/MODE.}
\usage{
project_multiomics(projectname, objects, fmeta = NULL)
}
\arguments{
\item{projectname}{Any string to name the project. All spaces and non-alphanumeric
characters will be removed to prevent issues with the visualizations. Required.}

\item{objects}{List of all omic projects or all pmart midpoints at the same tab (normalization or statistics).
Mixing of projects and midpoints is not allowed. Must contain 2-6 objects. There can be
no more than 2 metabolomics (1 of: NMR or GC/LC-MS), no more than 2 lipidomics datasets,
no more than 1 proteomics (peptide or protein) dataset, and no more than 1
transcriptomics dataset. Required.}

\item{fmeta}{Must be a dataframe or data table. If not provided, users can built it in
ipmart. Default is NULL.}
}
\value{
A multiomics project object
}
\description{
Constructs a multiomics project object from omic projects or pmart midpoints. Required.
}
\examples{
\dontrun{

library(pmartRdata)

# Generate midpoint with the examples in midpoint_pmart and save result as "midpoint"

# Make a metabolomics GC/LC MS project
metab_project <- project_omic(projectname = "My Metab Data",
                               datatype = "Metabolomics-GC/LC-MS",
                               edata = pmartRdata::metab_edata,
                               fdata = pmartRdata::metab_fdata,
                               edata_filename = "metab_edata",
                               fdata_filename = "metab_fdata")
                          
                                                    
# Make a lipidomics project
lipid_project <- project_omic(projectname = "My Lipid Data",
                               datatype = "Lipidomics-Positive",
                               edata = pmartRdata::lipid_pos_edata,
                               fdata = pmartRdata::lipid_pos_fdata,
                               edata_filename = "lipid_pos_edata",
                               fdata_filename = "lipid_pos_fdata")
                               
# Make a proteomics project
protein_project <- project_omic(projectname = "My Protein Data",
                                datatype = "Protein-level Label Free",
                                edata = pmartRdata::pro_edata,
                                fdata = pmartRdata::pro_fdata,
                                edata_filename = "pro_edata",
                                fdata_filename = "pro_fdata")
                                
                                
# Make an RNA seq project 
rna_project <- project_omic(projectname = "My RNA-Seq Data",
                             datatype = "RNA-seq",
                             edata = pmartRdata::rnaseq_edata,
                             fdata = pmartRdata::rnaseq_fdata,
                             edata_filename = "rnaseq_edata",
                             fdata_filename = "rnaseq_fdata")
                                
# Create an f_meta file
fmeta <- data.frame(
  "Proteins" = c(paste0("Mock", 1:3), paste0("Infection", c(1:4,6,7,9))),
  "Lipids" = c(paste0("Mock", 1:3), paste0("Infection", c(1:4,6,7,9))),
  "Metabolites" = c(paste0("Mock", 1:3), paste0("Infection", c(1:4,6,7,9)))
)
                          
# Finally, make the ipmart midpoint object, with an fmeta file
project_multiomics(projectname = "projects", objects = list(metab_project, lipid_project, protein_project), fmeta = fmeta)

# Make an ipmart midpoint object, but without fmeta. Note this specific example is for demo purposes only.
# In practice, all data should be from the same experiment.
project_multiomics(projectname = "project test", object = list(metab_project, lipid_project, protein_project, rna_project))


# Or use the pmart_midpoint examples 
project_multiomics(projectname = "midpoints", objects = list(pep_midpoint, lipid_midpoint))                          

}
}
