% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_constructor.R
\name{project_pmart}
\alias{project_pmart}
\title{Generate a project object to pass data from MAP to pmart}
\usage{
project_pmart(
  projectname,
  datatype,
  edata,
  fdata,
  emeta = NULL,
  edata_filename = NULL,
  fdata_filename = NULL,
  emeta_filename = NULL
)
}
\arguments{
\item{projectname}{Any string to name the project. All spaces and non-alphanumeric
characters will be removed to prevent issues with the visualizations. Required.}

\item{datatype}{Must be of the acceptable MAP omic class. See .is_datatype for list. Required.}

\item{edata}{Must be a dataframe or datatable. Required.}

\item{fdata}{Must be a dataframe or datatable. Required.}

\item{emeta}{Must be a dataframe or datatable. Optional.}

\item{edata_name}{The path or name for the edata file. Optional.}

\item{fdata_name}{The path or name for the fdata file. Optional.}

\item{emeta_name}{The path or name for the emeta file. Optional.}
}
\value{
A project pmart object
}
\description{
Constructs a project pmart object where edata and fdata are required.
}
\examples{
\dontrun{

library(pmartRdata)
project_pmart(projectname = "My Peptide Data",
             datatype = "Peptide-level Label Free",
             edata = pmartRdata::pep_edata,
             fdata = pmartRdata::pep_fdata,
             emeta = pmartRdata::pep_emeta,
             edata_filename = "pep_edata",
             fdata_filename = "pep_fdata",
             emeta_filename = "pep_emeta")

}
}
