% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tags_from_object.R
\name{pull_tags_from_object}
\alias{pull_tags_from_object}
\title{Reads the MAP object (project or midpoint) and generates appropriate tags}
\usage{
pull_tags_from_object(object)
}
\arguments{
\item{object}{A MAP object, either a project or midpoint. Required.}
}
\description{
Reads the MAP object (project or midpoint) and generates appropriate tags
}
\examples{
\dontrun{

library(pmartRdata)

# Make a pmart project object
project_pmart <- project_pmart(projectname = "My Peptide Data",
             datatype = "Peptide-level Label Free",
             edata = pmartRdata::pep_edata,
             fdata = pmartRdata::pep_fdata,
             emeta = pmartRdata::pep_emeta,
             edata_filename = "pep_edata",
             fdata_filename = "pep_fdata",
             emeta_filename = "pep_emeta")
pull_tags_from_object(project_pmart)
             
}

}
