% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_file.R
\name{put_file}
\alias{put_file}
\title{Send file to a moap_data_connection object and get back a UUID}
\usage{
put_file(con, filename, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}

\item{filename}{path to file}
}
\value{
UUIDthat can be used for retrieving the file (see \code{\link{get_file}})
}
\description{
Send file to a moap_data_connection object and get back a UUID
}
\seealso{
\code{\link{get_file}}, \code{\link{map_data_connection}}
}
