% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.R
\name{remove_data}
\alias{remove_data}
\title{Remove data object from a map_data_connection object using a UUID}
\usage{
remove_data(con, id, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}

\item{id}{ID returned by \code{\link{put_data}} or \code{\link{put_file}}}
}
\value{
TRUE or FALSE indicating success
}
\description{
Remove data object from a map_data_connection object using a UUID
}
