% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tags.R
\name{remove_tags}
\alias{remove_tags}
\title{Remove tags associated with an ID}
\usage{
remove_tags(con, id)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object using Min.io as the backend}

\item{id}{data object ID, typically returned by \code{\link{put_data}}}
}
\value{
nothing
}
\description{
Remove tags associated with an ID
}
\details{
Only valid when using Min.io as the backend storage mechanism.
}
\seealso{
\code{\link{get_tags}}, \code{\link{set_tags}}
}
