% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tags.R
\name{set_tags}
\alias{set_tags}
\title{Set tags associated with an ID}
\usage{
set_tags(con, id, tags)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object using Min.io as the backend}

\item{id}{data object ID, typically returned by \code{\link{put_data}}}

\item{tags}{list of key/value pairs}
}
\value{
nothing
}
\description{
Set tags associated with an ID
}
\details{
Only valid when using Min.io as the backend storage mechanism.
}
\seealso{
\code{\link{get_tags}}
}
