% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data_ids.R
\name{get_all_data_ids}
\alias{get_all_data_ids}
\title{Get all existing object IDs}
\usage{
get_all_data_ids(con, filter_tags = NULL, ...)
}
\arguments{
\item{con}{\code{\link{map_data_connection}} object}

\item{filter_tags}{(optional) Either a character vector or a named list/vector
with character elements.  If named, will return every object id with at least
one tag value that matches the corresponding value in the named list.
If unnamed, object ids with at least one tag key in filter_tags will be
returned.}
}
\value{
a character vector of IDs of existing data
}
\description{
Get all existing object IDs
}
\details{
If the SHINYPROXY_USERNAME environment variable is set this will
return all IDs for that user, otherwise it will return all IDs in the
data connection.
}
