% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetype_checks.R
\name{is_fmeta}
\alias{is_fmeta}
\title{Test if a file is an f_meta file}
\usage{
is_fmeta(edata_files, fmeta)
}
\arguments{
\item{edata_files}{A list of edata files. Required.}

\item{fmeta}{Must be a data.frame or data.table. Required.}
}
\value{
A boolean where TRUE means the file is an acceptable f_meta file.
}
\description{
Test if a file is an f_meta file
}
\examples{
\dontrun{

library(pmartRdata)
edata_files <- list(pmartRdata::lipid_edata, pmartRdata::pep_edata)
fmeta <- data.frame(
    PepSampleID = colnames(pmartRdata::pep_edata)[-1],
    LipidSampleID = c(paste0("Infection", 1:4), "", 
                      paste0("Infection", 6:9), paste0("Mock", 1:3))
)
is_fmeta(edata_files = edata_files, fmeta = fmeta)

}
}
