% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_constructor.R
\name{project_ipmart}
\alias{project_ipmart}
\title{Generate a project object to pass data from MAP to ipmart}
\usage{
project_ipmart(objects)
}
\arguments{
\item{objects}{List of projects and midpoints. Must contain 2-5 objects. There can be
no more than 2 metabolomics, no more than 2 lipidomics datasets, and no more than 1
proteomics (peptide or protein) dataset. Required.}
}
\value{
A project ipmart object
}
\description{
Constructs a project ipmart object from pmart projects and midpoints.
}
\examples{
\dontrun{

library(pmartRdata)

# Generate midpoint with the example in midpoint_pmart and save result as "midpoint"

# Make a metabolomics GC/LC MS project
metab_project <- project_pmart(projectname = "My Metab Data",
                               datatype = "Metabolomics-GC/LC-MS",
                               edata = pmartRdata::metab_edata,
                               fdata = pmartRdata::metab_fdata,
                               edata_filename = "metab_edata",
                               fdata_filename = "metab_fdata")
                               
# Make a metabolomics NMR project
nmr_project <- project_pmart(projectname = "My NMR Data",
                               datatype = "Metabolomics-NMR",
                               edata = pmartRdata::nmr_edata_identified,
                               fdata = pmartRdata::nmr_fdata_identified,
                               emeta = pmartRdata::nmr_emeta_identified,
                               edata_filename = "nmr_edata",
                               fdata_filename = "nmr_fdata",
                               emeta_filename = "nmr_emeta")
                          
                                                    
# Make a lipidomics project
lipid_project <- project_pmart(projectname = "My Lipid Data",
                               datatype = "Lipidomics-Positive",
                               edata = pmartRdata::lipid_edata,
                               fdata = pmartRdata::lipid_fdata,
                               edata_filename = "lipid_edata",
                               fdata_filename = "lipid_fdata")
                          
# Finally, make the ipmart midpoint object
project_ipmart(objects <- list(midpoint, metab_project, nmr_project, lipid_project))                          

}
}
